# docs/api/rest/openapi.yaml
openapi: 3.0.3
info:
  title: Travio API Gateway
  description: |
    Public-facing REST API for the Travio travel booking ecosystem.
    
    ## Conventions
    - **Casing**: All JSON fields use `snake_case` convention.
    - **Date Format**: `YYYY-MM-DD` for dates, RFC3339 for timestamps.
    
    ## Authentication
    All API requests (except Health, Discovery, and Search) require a valid JWT Access Token.
    Pass the token in the `Authorization` header: `Bearer <token>`.
    
    ## Errors
    Standard error format:
    ```json
    {
      "error": "description of error",
      "code": 400
    }
    ```
  version: 2.0.0
  contact:
    name: Travio Platform Team
    url: https://github.com/MuhibNayem/Travio
servers:
  - url: http://localhost:8888
    description: Local Development Server
...
paths:
...
  /v1/search/trips:
    get:
      summary: Search for available trips
      description: |
        Search for trips between two cities (Origin -> Destination).
        Note: Parameters denote City Names, not IDs.
      operationId: searchTrips
      tags: [Discovery]
      parameters:
        - name: from
          in: query
          description: Origin city name (e.g., "Dhaka")
          required: true
          schema: { type: string }
        - name: to
          in: query
          description: Destination city name (e.g., "Chittagong")
          required: true
          schema: { type: string }
        - name: date
          in: query
          description: Travel date (YYYY-MM-DD)
          schema: { type: string, format: date }
        - name: type
          in: query
          schema: { type: string, enum: [bus, train, ferry] }
      responses:
        '200':
          description: List of trips matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTripsResponse'
...
  /v1/stations:
    get:
      summary: List all stations
      operationId: listStations
      tags: [Discovery]
      parameters:
        - { name: city, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/PageParam'
      responses:
        '200':
          description: List of stations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStationsResponse'

  # ================= INVENTORY =================
  /v1/trips/{tripId}/availability:
    get:
      summary: Check real-time seat availability
      operationId: checkAvailability
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: tripId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: from
          in: query
          description: From Station ID
          required: true
          schema: { type: string, format: uuid }
        - name: to
          in: query
          description: To Station ID
          required: true
          schema: { type: string, format: uuid }
        - name: passengers
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckAvailabilityResponse' }
  /v1/trips/{tripId}/seatmap:
    get:
      summary: Get visualized seat map
      operationId: getSeatMap
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: tripId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: uuid }
        - name: to
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeatMap' }
  /v1/holds:
    post:
      summary: Hold seats temporarily
      description: Reserve seats for 10 minutes while user completes payment.
      operationId: holdSeats
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HoldSeatsRequest' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HoldSeatsResponse' }
  /v1/holds/{holdId}:
    delete:
      summary: Release seat hold
      operationId: releaseHold
      tags: [Inventory]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: holdId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Hold released

  # ================= ORDERS =================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 1 }
...
  schemas:
...
    SearchTripsResponse:
      type: object
      properties:
        results:
          type: array
          items: { $ref: '#/components/schemas/TripResult' }
        total: { type: integer }
        next_page: { type: string }
    TripResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        route_id: { type: string }
        departure_time: { type: string, format: date-time }
        arrival_time: { type: string, format: date-time }
        price: { type: integer, description: "Price in lowest currency unit" }
        operator: { type: string }
        available_seats: { type: integer }
        total_seats: { type: integer }
        type: { type: string, enum: [bus, train, ferry] }
        class: { type: string }
        from: { type: string }
        from_city: { type: string }
        to: { type: string }
        to_city: { type: string }
        duration: { type: integer, description: "Minutes" }
        distance: { type: integer, description: "KM" }
...
    StationResult:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        city: { type: string }
        state: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        amenities: { type: array, items: { type: string } }
    ListStationsResponse:
      type: object
      properties:
        stations:
          type: array
          items: { $ref: '#/components/schemas/StationResult' }
        total: { type: integer }

    # === INVENTORY ===
    CheckAvailabilityResponse:
      type: object
      properties:
        is_available: { type: boolean }
        available_seats: { type: integer }
        price_paisa: { type: integer }
        seats: 
          type: array
          items: { $ref: '#/components/schemas/SeatAvailability' }
    SeatAvailability:
      type: object
      properties:
        seat_id: { type: string }
        seat_number: { type: string }
        seat_class: { type: string }
        status: { type: string, enum: [available, held, booked] }
        seat_type: { type: string }
    SeatMap:
      type: object
      properties:
        rows: 
          type: array
          items: { type: object }
        legend: { type: object }
    HoldSeatsRequest:
      type: object
      required: [trip_id, seat_ids, from_station_id, to_station_id]
      properties:
        trip_id: { type: string }
        from_station_id: { type: string }
        to_station_id: { type: string }
        session_id: { type: string }
        seat_ids: 
          type: array
          items: { type: string }
    HoldSeatsResponse:
      type: object
      properties:
        hold_id: { type: string }
        success: { type: boolean }
        expires_at: { type: string, format: date-time }
        held_seat_ids: { type: array, items: { type: string } }
        failed_seat_ids: { type: array, items: { type: string } }
        failure_reason: { type: string }

    # === ORDERS ===
...